refactor: centralize blind signing logic and add safety checks
- Move duplicated blind signing functions from UI-specific files into a shared `menu_common.c`.
- Add NULL and length checks to `bytes_to_hex_display` to prevent potential crashes.
